[Move-only] Move LookupData struct and impl into lookup::prover
